Move API integration into a separate module
